tratamento de excecao para dados nao encontrados, com resposta 404 ao pesquisar, atualizar ou deletar

+++ src/main/java/br/com/clinica/app/adapters/service/MedicoServiceImpl.java
@@ -4,12 +4,12 @@ import br.com.clinica.app.adapters.database.entities.MedicoData;
 import br.com.clinica.app.adapters.database.repository.MedicoRepository;
 import br.com.clinica.app.adapters.presentation.response.MedicoResponse;
 import br.com.clinica.app.domain.entities.Medico;
+import br.com.clinica.app.domain.exception.DadoNaoEncontradoException;
 import br.com.clinica.app.domain.service.MedicoService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,7 +21,6 @@ public class MedicoServiceImpl implements MedicoService {
 
     private ModelMapper modelMapper = new ModelMapper();
     private final MedicoRepository medicoRepository;
-
 
     @Override
     public MedicoResponse create(Medico medico) {
@@ -35,28 +34,39 @@ public class MedicoServiceImpl implements MedicoService {
         List<MedicoData> medicosData = medicoRepository.findAll();
         List<MedicoResponse> medicosResponse = new ArrayList<>();
         for(MedicoData s : medicosData){
-            medicosResponse.add(modelMapper.map(s, MedicoResponse.class));
+            var medicoResponse = modelMapper.map(s, MedicoResponse.class);
+            medicosResponse.add(medicoResponse);
         }
         return medicosResponse;
     }
 
     @Override
-    public MedicoResponse update(Medico medico, Long id) {
-        MedicoData medicoData = modelMapper.map(medico, MedicoData.class);
+    public MedicoResponse update(Medico medico, Long id) throws DadoNaoEncontradoException {
+        var medicoPesquisado = medicoRepository.findById(id);
+        if(medicoPesquisado.isEmpty()){
+            throw new DadoNaoEncontradoException("medico nao encontrado pelo id: " + id);
+        }
+        var medicoData = modelMapper.map(medico, MedicoData.class);
         medicoData.setId(id);
         return modelMapper.map(medicoRepository.save(medicoData), MedicoResponse.class);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws DadoNaoEncontradoException {
+        var medicoData = medicoRepository.findById(id);
+        if(medicoData.isEmpty()){
+            throw new DadoNaoEncontradoException("medico nao encontrado pelo id: " + id);
+        }
         medicoRepository.deleteById(id);
     }
 
     @Override
-    public MedicoResponse getById(Long id) {
-        MedicoData secretarioData = medicoRepository.getById(id);
-        //TODO colocar excecao de not found aqui
-        return modelMapper.map(secretarioData, MedicoResponse.class);
+    public MedicoResponse getById(Long id) throws DadoNaoEncontradoException {
+        var medicoData = medicoRepository.findById(id);
+        if(medicoData.isEmpty()){
+            throw new DadoNaoEncontradoException("medico nao encontrado pelo id: " + id);
+        }
+        return modelMapper.map(medicoData, MedicoResponse.class);
     }
 
 }
