+++ src/main/java/br/com/clinica/app/adapters/service/ServicoServiceImpl.java
@@ -4,6 +4,7 @@ import br.com.clinica.app.adapters.database.entities.ServicoData;
 import br.com.clinica.app.adapters.database.repository.ServicoRepository;
 import br.com.clinica.app.adapters.presentation.response.ServicoResponse;
 import br.com.clinica.app.domain.entities.Servico;
+import br.com.clinica.app.domain.exception.DadoNaoEncontradoException;
 import br.com.clinica.app.domain.service.ServicoService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -40,21 +41,31 @@ public class ServicoServiceImpl implements ServicoService {
     }
 
     @Override
-    public ServicoResponse update(Servico servico, Long id) {
-        ServicoData servicoData = modelMapper.map(servico, ServicoData.class);
+    public ServicoResponse update(Servico servico, Long id) throws DadoNaoEncontradoException {
+        var servicoPesquisado = servicoRepository.findById(id);
+        if(servicoPesquisado.isEmpty()){
+            throw new DadoNaoEncontradoException("servico nao contrado pelo id: " + id);
+        }
+        var servicoData = modelMapper.map(servico, ServicoData.class);
         servicoData.setId(id);
         return modelMapper.map(servicoRepository.save(servicoData), ServicoResponse.class);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws DadoNaoEncontradoException {
+        var servicoData = servicoRepository.findById(id);
+        if(servicoData.isEmpty()){
+            throw new DadoNaoEncontradoException("servico nao contrado pelo id: " + id);
+        }
         servicoRepository.deleteById(id);
     }
 
     @Override
-    public ServicoResponse getById(Long id) {
-        ServicoData servicoData = servicoRepository.getById(id);
-        //TODO colocar excecao de not found aqui
+    public ServicoResponse getById(Long id) throws DadoNaoEncontradoException {
+        var servicoData = servicoRepository.findById(id);
+        if(servicoData.isEmpty()){
+            throw new DadoNaoEncontradoException("servico nao contrado pelo id: " + id);
+        }
         return modelMapper.map(servicoData, ServicoResponse.class);
     }
